TSM-5 - object model and daos * implemented spring's OpenViewInSessionInterceptor to support lazy loading inside of the view * Added EventFormFactory and EventForm.class * Updated Event test * Still need to combine Event and Feature

+++ trunk/tsm/src/test/java/com/tech4d/tsm/dao/IntegrationTestEventDao.java
@@ -1,18 +1,25 @@
 package com.tech4d.tsm.dao;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.List;
 
-import junit.framework.TestCase;
-
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.junit.BeforeClass;
+import org.junit.Test;
 import org.springframework.context.ApplicationContext;
+import org.springframework.orm.hibernate3.SessionFactoryUtils;
 
 import com.tech4d.tsm.model.Event;
 import com.tech4d.tsm.model.User;
 import com.tech4d.tsm.util.ContextUtil;
+import com.vividsolutions.jts.io.ParseException;
 
 /**
  * Test all dao functions
@@ -20,41 +27,58 @@ import com.tech4d.tsm.util.ContextUtil;
  * @author frank
  *
  */
-public class IntegrationTestEventDao extends TestCase {
+public class IntegrationTestEventDao  {
 
-    private EventDao eventDao;
+    private static EventDao eventDao;
     private static Event event = new Event();
-    
-    public void setUp() {
+    private static boolean initialized;
+
+    @BeforeClass
+    public static void setUp() {
         ApplicationContext appCtx = ContextUtil.getCtx();
         eventDao = (EventDao) appCtx.getBean("eventDao");
+        if (!initialized) {
+            initialized = true;
+            eventDao.deleteAll();
+            StyleUtil.setupStyle();
+        }
     }
     
-    /**
-     * Runs the first time
-     */
-    public void testInit() {
-        eventDao.deleteAll();        
-    }
-    
-    public void testSave() {
+    @Test
+    public void testSave() throws ParseException {
         event = createEvent("something is happening", new Date());
         Serializable id = eventDao.save(event);
         assertNotNull (id);
         assertEvents(1);
     }
     
-    public void findAll() {
+    @Test
+    public void testFindAll() {
         Collection<Event> events = eventDao.findAll();
         for (Event event : events) {
-//            System.out.println("event " + event.getId() + ", "  + event.getTitle());
+            System.out.println("event " + event.getId() + ", "  + event.getSourceUrl());
         }
         assertEquals(1, events.size());
+        
+        SessionFactory sessionFactory = eventDao.getSessionFactory();
+//        Session session = sessionFactory.openSession();
+//        Session session = SessionFactoryUtils.getSession(sessionFactory, true);
+//        session.beginTransaction();
+//        TransactionSynchronizationManager.bindResource(sessionFactory, new SessionHolder(session));
+
+        //Just for testing we will use OpenSessionInViewInterceptor for the real stuff
+        Session session = SessionFactoryUtils.getSession(sessionFactory, true);
+//        ThreadLocalSessionContext.bind(session);
+//        TransactionSynchronizationManager.bindResource(sessionFactory, new SessionHolder(session));
+
         for (Event event : events) {
+            session.refresh(event);
             assertEquals(2, event.getContributors().size());
         }
+        session.close();
     }
     
+    @Test
     public void testSaveOrUpdate() {
         String title = "a new title";
 //        event.setTitle(title);
@@ -62,17 +86,20 @@ public class IntegrationTestEventDao extends TestCase {
 //        assertEquals(title, eventDao.findById(event.getId()).getTitle());
     }
     
+    @Test
     public void testFindById() {
 //        assertEquals(event.getTitle(), eventDao.findById(event.getId()).getTitle());
     }
     
     
+    @Test
     public void testDelete() {
         eventDao.delete(event);
         assertEvents(0);
     }
     
-    public void testDeleteById() {
+    @Test
+    public void testDeleteById() throws ParseException {
         event = createEvent("something elese is happening", new Date());
         eventDao.save(event);
         assertEvents(1);
@@ -86,20 +113,22 @@ public class IntegrationTestEventDao extends TestCase {
         
     }
     
-    private Event createEvent(String title, Date date) {
+    private Event createEvent(String title, Date date) throws ParseException {
         Event event = new Event();
-//        event.setTitle(title);
-        Set<User> people = new HashSet<User>();
-        people.add(createPerson("Joe", 11));
-        people.add(createPerson("mary", 10));
+        List<User> people = new ArrayList<User>();
+        people.add(createUser("Joe", "1234"));
+        people.add(createUser("mary", "5678"));
         event.setContributors(people);
+        event.setFeature(FeatureUtil.createFeature());
+        event.setSourceUrl("http://www.wikipedia.com");
         return event;
     }
     
-    private User createPerson(String name, int age) {
+    
+    private User createUser(String username, String password) {
         User user = new User();
-//        user.setFirstname(name);
-//        user.setAge(age);
+        user.setUsername(username);
+        user.setPassword(password);
         return user;
     }
